+++ src/Accounts/Authentication/Factories/AuthenticationFactory.cs
@@ -482,10 +482,10 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
         }
 
         /// <summary>
-        /// Remove a user from token cache.
+        /// Remove any stored credentials for the given user and the Azure environment used.
         /// </summary>
-        /// <param name="account"></param>
-        /// <param name="environment"></param>
+        /// <param name="account">The account to remove credentials for</param>
+        /// <param name="environment">The environment which account belongs to</param>
         public void RemoveUser(IAzureAccount account, IAzureEnvironment environment)
         {
             if (account != null && !string.IsNullOrEmpty(account.Id) && !string.IsNullOrWhiteSpace(account.Type))
@@ -513,10 +513,10 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
                             // make best effort to remove credentials
                         }
 
-                        RemoveFromTokenCache(account);
+                        RemoveFromTokenCache(account, environment?.ActiveDirectoryAuthority);
                         break;
                     case AzureAccount.AccountType.User:
-                        RemoveFromTokenCache(account);
+                        RemoveFromTokenCache(account, environment?.ActiveDirectoryAuthority);
                         break;
                 }
             }
@@ -558,7 +558,7 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
             return account.GetProperty(tokenKey);
         }
 
-        private void RemoveFromTokenCache(IAzureAccount account)
+        private void RemoveFromTokenCache(IAzureAccount account, string authority = null)
         {
             PowerShellTokenCacheProvider tokenCacheProvider;
             if (!AzureSession.Instance.TryGetComponent(PowerShellTokenCacheProvider.PowerShellTokenCacheProviderKey, out tokenCacheProvider))
@@ -566,12 +566,12 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
                 throw new NullReferenceException(Resources.AuthenticationClientFactoryNotRegistered);
             }
 
-            var publicClient = tokenCacheProvider.CreatePublicClient();
+            var publicClient = tokenCacheProvider.CreatePublicClient(authority);
             var accounts = publicClient.GetAccountsAsync()
                             .ConfigureAwait(false).GetAwaiter().GetResult();
             var tokenAccounts = accounts.Where(a => MatchCacheItem(account, a));
             foreach (var tokenAccount in tokenAccounts)
-                {
+            {
                 publicClient.RemoveAsync(tokenAccount)
                                 .ConfigureAwait(false).GetAwaiter().GetResult();
             }