+++ src/MobileNetwork/MobileNetwork.Autorest/custom/New-AzMobileNetworkSite.ps1
@@ -25,12 +25,13 @@ Must be created in the same location as its parent mobile network.
 New-AzMobileNetworkSite -MobileNetworkName azps-mn -Name azps-mn-site -ResourceGroupName azps_test_group -Location eastus -Tag @{"site"="123"}
 
 .Outputs
-Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Models.Api20221101.ISite
+Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Models.ISite
 .Link
 https://learn.microsoft.com/powershell/module/az.mobilenetwork/new-azmobilenetworksite
 #>
 function New-AzMobileNetworkSite {
-    [OutputType([Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Models.Api20221101.ISite])]
+    [OutputType([Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Models.ISite])]
+    [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Runtime.CmdletBreakingChange("14.5.0", "0.5.0", "2025/09/30")]
     [CmdletBinding(DefaultParameterSetName = 'CreateExpanded', PositionalBinding = $false, SupportsShouldProcess, ConfirmImpact = 'Medium')]
     param(
         [Parameter(Mandatory)]
@@ -68,16 +69,16 @@ function New-AzMobileNetworkSite {
 
         [Parameter()]
         [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Category('Body')]
-        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Runtime.Info(PossibleTypes = ([Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Models.Api30.ITrackedResourceTags]))]
+        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Runtime.Info(PossibleTypes = ([Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Models.ITrackedResourceTags]))]
         [System.Collections.Hashtable]
         # Resource tags.
         ${Tag},
 
         # New-AzMobileNetworkPacketCoreControlPlane
         [Parameter()]
-        [ArgumentCompleter([Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.PlatformType])]
+        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.PSArgumentCompleterAttribute("AKS-HCI", "3P-AZURE-STACK-HCI")]
         [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Category('Body')]
-        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.PlatformType]
+        [System.String]
         # The platform type where packet core is deployed.
         ${PlatformType},
 
@@ -119,23 +120,23 @@ function New-AzMobileNetworkSite {
         ${AzureStackEdgeDeviceId},
 
         [Parameter()]
-        [ArgumentCompleter([Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.AuthenticationType])]
+        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.PSArgumentCompleterAttribute("AAD", "Password")]
         [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Category('Body')]
-        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.AuthenticationType]
+        [System.String]
         # How to authenticate users who access local diagnostics APIs.
         ${LocalDiagnosticAccessAuthenticationType},
 
         [Parameter()]
-        [ArgumentCompleter([Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.CoreNetworkType])]
+        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.PSArgumentCompleterAttribute("5GC", "EPC")]
         [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Category('Body')]
-        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.CoreNetworkType]
+        [System.String]
         # The core network technology generation (5G core or EPC / 4G core).
         ${CoreNetworkTechnology},
 
         [Parameter()]
-        [ArgumentCompleter([Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.BillingSku])]
+        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.PSArgumentCompleterAttribute("G0", "G1", "G2", "G5", "G10")]
         [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Category('Body')]
-        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.BillingSku]
+        [System.String]
         # The SKU defining the throughput and SIM allowances for this packet core control plane deployment.
         ${Sku},
 
@@ -175,9 +176,9 @@ function New-AzMobileNetworkSite {
         ${DnsAddress},
 
         [Parameter()]
-        [ArgumentCompleter([Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.NaptEnabled])]
+        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.PSArgumentCompleterAttribute("Enabled", "Disabled")]
         [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Category('Body')]
-        [Microsoft.Azure.PowerShell.Cmdlets.MobileNetwork.Support.NaptEnabled]
+        [System.String]
         # Whether NAPT is enabled for connections to this attached data network.
         ${NaptConfigurationEnabled},