+++ src/NetAppFiles/NetAppFiles/Models/PSNetAppFilesBackup.cs
@@ -30,7 +30,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Models
         /// Gets or sets Resource location
         /// </summary>
         public string Location { get; set; }
-        
+
         /// <summary>
         /// Gets resource Id
         /// </summary>
@@ -113,5 +113,13 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Models
         ///  The name of the snapshot
         /// </remarks>        
         public string SnapshotName { get; set; }
+
+        /// <summary>
+        /// Gets IsLargeVolume 
+        /// </summary>
+        /// <remarks>
+        /// Specifies if the backup is for a large volume
+        /// </remarks>
+        public bool? IsLargeVolume { get; set; }
     }
 }