+++ src/RecoveryServices/RecoveryServices.Backup.Test/ScenarioTests/AzureWorkload/ContainerTests.ps1
@@ -16,6 +16,40 @@ $containerName = "psbvtsqlvm"
 $resourceGroupName = "pstestwlRG1bca8"
 $vaultName = "pstestwlRSV1bca8"
 $resourceId = "/subscriptions/38304e13-357e-405e-9e9a-220351dcce8c/resourceGroups/pscloudtestrg/providers/Microsoft.Compute/virtualMachines/psbvtsqlvm"
+$sqlRegTestVlt = "utkvlt"
+$sqlRegTestVm = "utkvm"
+$sqlRegTestRg1 = "rg1"
+$sqlRegTestRg2 = "rg2"
+$sqlRegTestVM2Id = "/subscriptions/af95aa3c-30fd-41c6-a938-4b3676fc36fb/resourceGroups/rg2/providers/Microsoft.Compute/virtualMachines/utkvm"
+
+function Test-SQLContainerRegError
+{
+    $vault1 = Get-AzRecoveryServicesVault -ResourceGroupName $sqlRegTestRg1 -Name $sqlRegTestVlt
+    $vault2 = Get-AzRecoveryServicesVault -ResourceGroupName $sqlRegTestRg2 -Name $sqlRegTestVlt
+
+   #$Unregister containers if already registered
+   Get-AzRecoveryServicesBackupContainer `
+         -VaultId $vault1.ID `
+         -ContainerType AzureVMAppContainer `
+         -FriendlyName $sqlRegTestVm | Unregister-AzRecoveryServicesBackupContainer -VaultId $vault1.ID -Force
+
+   Get-AzRecoveryServicesBackupContainer `
+         -VaultId $vault2.ID `
+         -ContainerType AzureVMAppContainer `
+         -FriendlyName $sqlRegTestVm | Unregister-AzRecoveryServicesBackupContainer -VaultId $vault2.ID -Force
+
+    $v = get-azrecoveryservicesvault -ResourceGroupName $sqlRegTestRg2 -Name $sqlRegTestVlt    
+    Set-AzRecoveryServicesVaultContext -Vault $v
+
+    Register-AzRecoveryServicesBackupContainer -ResourceId $sqlRegTestVM2Id -VaultId $v.Id -WorkloadType "MSSQL" -BackupManagementType "AzureWorkload" -Force
+
+    $container = Get-AzRecoveryServicesBackupContainer `
+         -VaultId $vault2.ID `
+         -ContainerType AzureVMAppContainer `
+         -FriendlyName $sqlRegTestVm
+
+    Assert-True { $container.FriendlyName -eq $sqlRegTestVm }
+}
 
 function Test-AzureVmWorkloadUnDeleteContainer
 {
@@ -78,7 +112,7 @@ function Get-AzureVmWorkloadContainer
       $containers = Get-AzRecoveryServicesBackupContainer `
          -VaultId $vault.ID `
          -ContainerType AzureVMAppContainer;
-      Assert-True { $containers[0].FriendlyName -contains $containerName }
+      Assert-True { $containers[1].FriendlyName -contains $containerName }
 
       # VARIATION-2: Get Containers with friendly name filter
       $containers = Get-AzRecoveryServicesBackupContainer `
@@ -92,7 +126,7 @@ function Get-AzureVmWorkloadContainer
          -VaultId $vault.ID `
          -ContainerType AzureVMAppContainer `
          -ResourceGroupName $resourceGroupName;
-      Assert-True { $containers[0].FriendlyName -contains $containerName }
+      Assert-True { $containers[1].FriendlyName -contains $containerName }
    
       # VARIATION-4: Get Containers with friendly name and resource group filters
       $containers = Get-AzRecoveryServicesBackupContainer `