+++ src/Accounts/Accounts/ChangeLog.md
@@ -22,6 +22,7 @@
 
 ## Version 5.0.2
 * Upgrade Azure.Core to 1.45.0
+* Added definition of Bleu cloud so it can be used with commands like `Enter-AzVM`
 
 ## Version 5.0.1
 * Force `Get-AzAccessToken` to always return SecureString as long as `AsSecureString` is set